rypter/requirements.txt
cryptography>=41.0.0
pyinstaller
python-docx>=0.8.11
pillow>=10.0.0
pefile>=2023.2.7
requests
colorama
